--- src/vui/components/toggle/Toggle.d.ts
@@ -1,8 +1,0 @@
-type Props = {
-    id?: string;
-    checked?: boolean;
-    onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-    label?: string;
-};
-export declare const VuiToggle: ({ id, checked, onChange, label, ...rest }: Props) => import("react/jsx-runtime").JSX.Element;
-export {};
